+++ CHANGELOG.rst
@@ -15,6 +15,38 @@ ProPlot v1.0.0 (2020-##-##)
 This will be published when some major refactoring tasks are completed.
 See :pr:`45`, :pr:`46`, and :pr:`50`.
 
+ProPlot v0.2.0 (2019-12-02)
+===========================
+Features
+--------
+- Support manual resizing for all backends, including ``osx`` and ``qt``.
+
+Bug fixes
+---------
+- Disable automatic resizing for the ``nbAgg`` interactive inline backend. Found no
+  suitable workaround.
+- Remove several `~matplotlib.rcParams` settings from the default ``.proplotrc``.
+  ``savefig.dpi`` and ``savefig.transparent`` were affecting the ``nbAgg``
+  interactive inline backend.
+
+Deprecated
+----------
+- Remove the ``nbsetup`` rc setting in favor of separate ``autosave``, ``autoreload``,
+  and ``matplotlib`` settings for triggering the respective ``%`` magic commands.
+  (``nbsetup`` is still accepted but no longer documented).
+- Rename the ``format`` rc setting in favor of the ``inlinefmt`` setting
+  (``format`` is still accepted but no longer documented).
+- Rename ``FlexibleGridSpec`` and ``FlexibleSubplotSpec`` to ``GridSpec``
+  and ``SubplotSpec`` (until :pr:`50` is merged it is impossible
+  to use these manually, so this won't bother anyone).
+
+Internals
+---------
+- Rename ``rcParamsCustom`` to ``rcParamsLong`` (this is inaccessible to the user).
+- When calling ``fig.canvas.print_figure()`` on a stale figure, call ``fig.canvas.draw()``
+  first. May be overkill for `~matplotlib.figure.Figure.savefig` but critical for
+  correctly displaying already-drawn notebook figures.
+
 ProPlot v0.1.3 (2019-12-01)
 ===========================
 Internals